parse targets count field as int

+++ web/static/chaos-game-3d.js
@@ -247,7 +247,7 @@ function resetTargetsLocations(verticesCount) {
 }
 
 function initializeTargets() {
-  resetTargetsLocations(parseFloat(document.getElementById('vertices').value));
+  resetTargetsLocations(parseInt(document.getElementById('vertices').value));
   createTargetMeshes();
 }
